+++ src/lib/contacts/contacts.ts
@@ -2,6 +2,8 @@ import { Incident } from "incident";
 import { UnexpectedHttpStatusError } from "../errors/http";
 import { Context } from "../interfaces/api/context";
 import * as io from "../interfaces/http-io";
+import { JSON_READER } from "../json-reader";
+import { MriKey } from "../mri";
 import { Contact } from "../types/contact";
 import { Invite } from "../types/invite";
 import { Url } from "../types/url";
@@ -9,7 +11,7 @@ import { getContacts } from "./api/get-contacts";
 import { $GetInvitesResult, GetInvitesResult } from "./api/get-invites";
 import * as contactsUrl from "./contacts-url";
 
-export interface ContactsInterface {
+export interface ContactServiceInterface {
   /**
    * Get the pending incoming contact invitations.
    *
@@ -25,12 +27,15 @@ export interface ContactsInterface {
    * @return The list of contacts.
    */
   getContacts(apiContext: Context): Promise<Contact[]>;
+
+  getContactById(apiContext: Context, contactId: MriKey): Promise<Contact>;
 }
 
 /**
  * @internal
  */
-export class ContactsService {
+export class ContactService implements ContactServiceInterface {
+
   private readonly httpIo: io.HttpIo;
 
   constructor(httpIo: io.HttpIo) {
@@ -50,23 +55,20 @@ export class ContactsService {
     if (response.statusCode !== 200) {
       UnexpectedHttpStatusError.create(response, new Set([200]), request);
     }
-    let parsed: any;
-    try {
-      parsed = JSON.parse(response.body);
-    } catch (err) {
-      throw new Incident(err, "UnexpectedResponseBody", {body: response.body});
-    }
-
     let result: GetInvitesResult;
     try {
-      result = $GetInvitesResult.readJson(parsed);
+      result = $GetInvitesResult.read(JSON_READER, response.body);
     } catch (err) {
-      throw new Incident(err, "UnexpectedResult", {body: parsed});
+      throw new Incident(err, "UnexpectedResponseBody", {body: response.body});
     }
     return result.inviteList;
   }
 
   async getContacts(apiContext: Context): Promise<Contact[]> {
     return getContacts(this.httpIo, apiContext);
+  }
+
+  async getContactById(apiContext: Context, contactId: MriKey): Promise<Contact> {
+    throw new Incident("NotImplemented", "getContactById");
   }
 }