+++ src/lib/interfaces/api/events.ts
@@ -1,4 +1,3 @@
-import { ParsedConversationId } from "./api";
 import { Resource } from "./resources";
 
 export interface EventMessage {