+++ src/lib/types/api-profile.ts
@@ -1,13 +1,13 @@
-import { VersionedType } from "kryo/types";
+import { IoType } from "kryo/types";
+import { AnyType } from "kryo/types/any";
 import { ArrayType } from "kryo/types/array";
-import { DocumentType } from "kryo/types/document";
-import { JsonType } from "kryo/types/json";
+import { DocumentIoType, DocumentType } from "kryo/types/document";
 import { NullType } from "kryo/types/null";
 import { TryUnionType } from "kryo/types/try-union";
 import { Ucs2StringType } from "kryo/types/ucs2-string";
 import { $Url, Url } from "./url";
 
-function nullable(type: VersionedType<any, any, any, any>): VersionedType<any, any, any, any> {
+function nullable(type: IoType<any>): IoType<any> {
   return new TryUnionType({
     variants: [
       new NullType(),
@@ -95,27 +95,26 @@ export interface ApiProfile {
   username: string;
 }
 
-export const $ApiProfile: DocumentType<ApiProfile> = new DocumentType<ApiProfile>({
+export const $ApiProfile: DocumentIoType<ApiProfile> = new DocumentType<ApiProfile>({
   properties: {
     firstname: {type: new Ucs2StringType({maxLength: Infinity})},
     lastname: {type: nullable(new Ucs2StringType({maxLength: Infinity}))},
-    birthday: {type: new JsonType()},
-    gender: {type: new JsonType()},
-    language: {type: new JsonType()},
-    country: {type: new JsonType()},
-    province: {type: new JsonType()},
-    city: {type: new JsonType()},
-    homepage: {type: new JsonType()},
-    about: {type: new JsonType()},
+    birthday: {type: new AnyType()},
+    gender: {type: new AnyType()},
+    language: {type: new AnyType()},
+    country: {type: new AnyType()},
+    province: {type: new AnyType()},
+    city: {type: new AnyType()},
+    homepage: {type: new AnyType()},
+    about: {type: new AnyType()},
     emails: {type: new ArrayType({itemType: new Ucs2StringType({maxLength: Infinity}), maxLength: Infinity})},
-    jobtitle: {type: new JsonType()},
-    phoneMobile: {type: new JsonType()},
-    phoneHome: {type: new JsonType()},
-    phoneOffice: {type: new JsonType()},
-    mood: {type: new JsonType()},
-    richMood: {type: new JsonType()},
+    jobtitle: {type: new AnyType()},
+    phoneMobile: {type: new AnyType()},
+    phoneHome: {type: new AnyType()},
+    phoneOffice: {type: new AnyType()},
+    mood: {type: new AnyType()},
+    richMood: {type: new AnyType()},
     avatarUrl: {type: nullable($Url)},
     username: {type: new Ucs2StringType({maxLength: Infinity})},
   },
-  ignoreExtraKeys: true,
 });