+++ src/lib/types/contact-group.ts
@@ -1,6 +1,6 @@
+import { $Boolean } from "kryo/builtins/boolean";
 import { CaseStyle } from "kryo/case-style";
-import { BooleanType } from "kryo/types/boolean";
-import { DocumentType } from "kryo/types/document";
+import { DocumentIoType, DocumentType } from "kryo/types/document";
 import { Ucs2StringType } from "kryo/types/ucs2-string";
 
 /**
@@ -19,12 +19,11 @@ export interface ContactGroup {
   isFavorite?: boolean;
 }
 
-export const $ContactGroup: DocumentType<ContactGroup> = new DocumentType<ContactGroup>({
+export const $ContactGroup: DocumentIoType<ContactGroup> = new DocumentType<ContactGroup>({
   properties: {
     id: {type: new Ucs2StringType({maxLength: Infinity})},
     name: {type: new Ucs2StringType({maxLength: Infinity})},
-    isFavorite: {type: new BooleanType(), optional: true},
+    isFavorite: {type: $Boolean, optional: true},
   },
-  rename: CaseStyle.SnakeCase,
-  ignoreExtraKeys: true,
+  changeCase: CaseStyle.SnakeCase,
 });