+++ src/lib/types/empty-object.ts
@@ -1,0 +1,8 @@
+import { DocumentType } from "kryo/types/document";
+
+export interface EmptyObject {
+}
+
+export const $EmptyObject: DocumentType<EmptyObject> = new DocumentType<EmptyObject>({
+  properties: {},
+});