+++ src/lib/types/message-type.ts
@@ -1,0 +1,11 @@
+import { CaseStyle } from "kryo/case-style";
+import { TsEnumType } from "kryo/types/ts-enum";
+
+export enum MessageType {
+  RichText,
+}
+
+export const $MessageType: TsEnumType<MessageType> = new TsEnumType<MessageType>({
+  enum: MessageType,
+  changeCase: CaseStyle.PascalCase,
+});